♻️ refactor: rename internal module name
