+++ playing_cards/__init__.py
@@ -2,6 +2,6 @@
 Playing cards from the French-suited, standard 52-card pack.
 """
 
-from playing_cards.deck import Card, Colour, Deck, Rank, Suit
+from playing_cards.models import Card, Colour, Deck, Rank, Suit
 
 __all__ = ["Card", "Colour", "Deck", "Rank", "Suit"]
